Extract keymap.c from misc2.c
