+++ src/eval.c
@@ -34,6 +34,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "keymap.h"
 #include "file_search.h"
 #include "garray.h"
 #include "move.h"
