+++ src/option.c
@@ -53,6 +53,7 @@
 #include "message.h"
 #include "misc1.h"
 #include "misc2.h"
+#include "keymap.h"
 #include "crypt.h"
 #include "garray.h"
 #include "move.h"
