+++ src/term.c
@@ -35,6 +35,7 @@
 #include "message.h"
 #include "misc2.h"
 #include "garray.h"
+#include "keymap.h"
 #include "move.h"
 #include "normal.h"
 #include "option.h"
